fixes URL resource delete
phpcs
