+++ plugin/announcement/Crud/AnnouncementCrud.php
@@ -37,7 +37,7 @@ class AnnouncementCrud
     }
 
     /**
-     * @param CreateEvent $event
+     * @param DeleteEvent $event
      */
     public function preDelete(DeleteEvent $event)
     {
