+++ plugin/forum/Resources/config/services/listener.yml
@@ -21,7 +21,6 @@ services:
             - '@security.token_storage'
         tags:
             - { name: kernel.event_listener, event: resource.claroline_forum.load, method: onOpen }
-            - { name: kernel.event_listener, event: resource.claroline_forum.delete, method: onDelete }
             - { name: kernel.event_listener, event: transfer.claroline_forum.export, method: onExport }
             - { name: kernel.event_listener, event: transfer.claroline_forum.import.before, method: onImportBefore }
             - { name: kernel.event_listener, event: transfer.claroline_forum.import.after, method: onImport }
