+++ plugin/lesson/Resources/config/services/listener.yml
@@ -16,7 +16,6 @@ services:
             - { name: kernel.event_listener, event: resource.icap_lesson.copy, method: onCopy }
             - { name: kernel.event_listener, event: transfer.icap_lesson.export, method: onExport }
             - { name: kernel.event_listener, event: transfer.icap_lesson.import.after, method: onImport }
-            - { name: kernel.event_listener, event: resource.icap_lesson.delete, method: onDelete }
 
     icap.listener.lesson_log_listener:
         class: Icap\LessonBundle\Listener\LogListener
