+++ plugin/path/Resources/config/services/listener.yml
@@ -14,6 +14,5 @@ services:
            - '@Innova\PathBundle\Manager\UserProgressionManager'
         tags:
             - { name: kernel.event_listener, event: resource.innova_path.load, method: onLoad }
-            - { name: kernel.event_listener, event: resource.innova_path.delete, method: onDelete }
             - { name: kernel.event_listener, event: resource.innova_path.copy, method: onCopy }
             - { name: kernel.event_listener, event: resource_evaluation, method: onEvaluation }
