+++ plugin/rss/Resources/config/services/services.yml
@@ -12,4 +12,3 @@ services:
            - '@Claroline\AppBundle\API\SerializerProvider'
         tags:
             - { name: kernel.event_listener, event: resource.rss_feed.load, method: onLoad }
-            - { name: kernel.event_listener, event: resource.rss_feed.delete, method: onDelete }
